requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-ssh
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-microsoft-psrp